feat(redis/run.sh) make it easier to bootstrap a slave or sentinel node
Previously, we'd need an existing sentinel cluster for either slave or sentinel instances to boot up.

This commit introduces a fix where we can specify either a fallback IP address of a (potentially) known master instance,
via the file /tmp/REDIS_MASTER_HOST, or the environment variable REDIS_MASTER_HOST. We can either set the variable at
run time, or write to the file after exec'ing into the container after it has started. Either way, this allows the instance
to find a Redis master, and not get stuck in a boot loop